[UPDATE] 레벨 변화 테스트 완료

+++ Assets/Scripts/UI/PlayerLevelUI.cs
@@ -54,7 +54,7 @@ public class PlayerLevelUI : MonoBehaviour
     private void UpdateExpUI()
     {
         expText.text = $"{currentExp} / {maxExp}";
-        expBar.value = currentExp / maxExp;
+        expBar.value = (float)currentExp / maxExp;
     }
 
     private void UpdateLevel(int level)
